requirements.txt
chatterbox-tts
runpod
torchaudio
numpy
scipy